use feature for gateway
