+++ zkstack_cli/crates/zkstack/src/commands/chain/utils.rs
@@ -1,0 +1,13 @@
+use ethers::abi::encode;
+use zksync_types::{web3::keccak256, Address, H256, L2_NATIVE_TOKEN_VAULT_ADDRESS, U256};
+
+
+pub fn encode_ntv_asset_id(l1_chain_id: U256, addr: Address) -> H256 {
+    let encoded_data = encode(&[
+        ethers::abi::Token::Uint(l1_chain_id),
+        ethers::abi::Token::Address(L2_NATIVE_TOKEN_VAULT_ADDRESS),
+        ethers::abi::Token::Address(addr),
+    ]);
+
+    H256(keccak256(&encoded_data))
+}
